halfway-fix-sign-up-bug
