+++ src/App.js
@@ -1,9 +1,10 @@
 import React from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import Home from "./pages/Home";
-import SignUp from "./pages/SignUp";
-import Error from "./pages/Error";
-import SignIn from "./pages/SignIn";
+import HomePage from "./pages/HomePage";
+import SignUpPage from "./pages/SignUpPage";
+import AccountPage from "./pages/AccountPage";
+import ErrorPage from "./pages/ErrorPage";
+import SignInPage from "./pages/SignInPage";
 import { useState, useEffect } from "react";
 import { supabase } from "./supabaseClient";
 
@@ -21,11 +22,18 @@ function App() {
     return (
         <Router>
             <Routes>
-                <Route path="/" element={<Home />} />
-                <Route path="/home" element={<Home />} />
-                <Route path="/signin" element={<SignIn session={session} />} />
-                <Route path="/signup" element={<SignUp />} />
-                <Route path="*" element={<Error />} />
+                <Route path="/" element={<HomePage />} />
+                <Route path="/home" element={<HomePage />} />
+                <Route
+                    path="/signin"
+                    element={<SignInPage session={session} />}
+                />
+                <Route path="/signup" element={<SignUpPage />} />
+                <Route
+                    path="/account"
+                    element={<AccountPage session={session} />}
+                />
+                <Route path="*" element={<ErrorPage />} />
             </Routes>
         </Router>
     );
