+++ src/pages/ErrorPage.js
@@ -1,7 +1,7 @@
 import React from "react";
 
-const Error = () => {
+const ErrorPage = () => {
     return <div>Error: Page Not Found!</div>;
 };
 
-export default Error;
+export default ErrorPage;
