--- src/pages/SignUp.js
@@ -1,12 +1,0 @@
-import React from "react";
-import SignUpPage from "../components/SignUpPage";
-
-const SignUp = () => {
-    return (
-        <>
-            <SignUpPage></SignUpPage>
-        </>
-    );
-};
-
-export default SignUp;
