+++ src/pages/SignUpPage.js
@@ -1,0 +1,12 @@
+import React from "react";
+import SignUp from "../components/SignUp";
+
+const SignUpPage = () => {
+    return (
+        <>
+            <SignUp />
+        </>
+    );
+};
+
+export default SignUpPage;
